# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CatalogItem {
  id: number;
  name: string;
  category: string;
  packageSize: string;
  communityPrice: number;
  retailPrice: number;
  savingsPercent: number;
  stockQuantity: number;
  imageKey: string;
  popular?: boolean;
}

export interface PricingSummary {
  lastCalculatedAt: string;
  globalOffsetPercent: number;
  totalItems: number;
  projectedCommunitySavings: number;
  status: string;
}

export interface PricingInput {
  /**
     * @minimum 0
     * @maximum 50
     */
  offsetPercent: number;
}

export interface Hub {
  id: number;
  name: string;
  region: string;
  status: string;
  activeOrders: number;
  stockHealth: number;
  agentCount: number;
}

export interface WardZone {
  id: number;
  name: string;
  municipality: string;
  hubName: string;
  households: number;
  status: string;
}

export interface OrderLine {
  itemName: string;
  packageSize: string;
  quantity: number;
  unitPrice: number;
}

export interface Order {
  id: number;
  reference: string;
  clientName: string;
  clientPhone: string;
  hubName: string;
  address: string;
  orderSource: string;
  status: string;
  totalAmount: number;
  itemCount: number;
  createdAt: string;
  lines: OrderLine[];
}

export interface OrderInput {
  clientName: string;
  clientPhone: string;
  hubId: number;
  address: string;
  orderSource: string;
  /** @minItems 1 */
  lines: OrderLine[];
}

export interface OrderStatusInput {
  status: string;
}

export interface WhatsAppOrderInput {
  message: string;
  clientName: string;
  clientPhone: string;
}

export interface WhatsAppOrderResult {
  order: Order;
  parsedMessage: string;
  confidence: number;
}

export interface Delivery {
  id: number;
  orderReference: string;
  agentName: string;
  agentInitials: string;
  pickup: string;
  dropoff: string;
  status: string;
  eta: string;
}

export interface DeliveryStatusInput {
  status: string;
}

export interface TrendPoint {
  label: string;
  orders: number;
  retained: number;
}

export interface ForecastPoint {
  item: string;
  daysRemaining: number;
  confidence: number;
  trend: string;
}

export interface ActivityItem {
  id: number;
  title: string;
  detail: string;
  timestamp: string;
  tone: string;
}

export interface DashboardSummary {
  totalOrders: number;
  activeHubs: number;
  localSavings: number;
  retainedCurrency: number;
  corporateLeakage: number;
  serviceReinvestment: number;
  orderTrend: TrendPoint[];
  demandForecast: ForecastPoint[];
  recentActivity: ActivityItem[];
}

export type ListOrdersParams = {
status?: string;
};