# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActivityItem {
  id: number;
  title: string;
  detail: string;
  timestamp: string;
  tone: string;
}