# lib/api-zod/src/generated/types/catalogItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogItem {
  id: number;
  name: string;
  category: string;
  packageSize: string;
  communityPrice: number;
  retailPrice: number;
  savingsPercent: number;
  stockQuantity: number;
  imageKey: string;
  popular?: boolean;
}