# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from './activityItem';
import type { ForecastPoint } from './forecastPoint';
import type { TrendPoint } from './trendPoint';

export interface DashboardSummary {
  totalOrders: number;
  activeHubs: number;
  localSavings: number;
  retainedCurrency: number;
  corporateLeakage: number;
  serviceReinvestment: number;
  orderTrend: TrendPoint[];
  demandForecast: ForecastPoint[];
  recentActivity: ActivityItem[];
}