# lib/api-zod/src/generated/types/delivery.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Delivery {
  id: number;
  orderReference: string;
  agentName: string;
  agentInitials: string;
  pickup: string;
  dropoff: string;
  status: string;
  eta: string;
}