# lib/api-zod/src/generated/types/forecastPoint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ForecastPoint {
  item: string;
  daysRemaining: number;
  confidence: number;
  trend: string;
}