# lib/api-zod/src/generated/types/hub.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Hub {
  id: number;
  name: string;
  region: string;
  status: string;
  activeOrders: number;
  stockHealth: number;
  agentCount: number;
}