# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderLine } from './orderLine';

export interface Order {
  id: number;
  reference: string;
  clientName: string;
  clientPhone: string;
  hubName: string;
  address: string;
  orderSource: string;
  status: string;
  totalAmount: number;
  itemCount: number;
  createdAt: string;
  lines: OrderLine[];
}