# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderLine } from './orderLine';

export interface OrderInput {
  clientName: string;
  clientPhone: string;
  hubId: number;
  address: string;
  orderSource: string;
  /** @minItems 1 */
  lines: OrderLine[];
}