# lib/api-zod/src/generated/types/orderLine.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderLine {
  itemName: string;
  packageSize: string;
  quantity: number;
  unitPrice: number;
}