# lib/api-zod/src/generated/types/pricingInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PricingInput {
  /**
     * @minimum 0
     * @maximum 50
     */
  offsetPercent: number;
}