# lib/api-zod/src/generated/types/pricingSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PricingSummary {
  lastCalculatedAt: string;
  globalOffsetPercent: number;
  totalItems: number;
  projectedCommunitySavings: number;
  status: string;
}