# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  label: string;
  orders: number;
  retained: number;
}