# lib/api-zod/src/generated/types/wardZone.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WardZone {
  id: number;
  name: string;
  municipality: string;
  hubName: string;
  households: number;
  status: string;
}