# lib/api-zod/src/generated/types/whatsAppOrderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WhatsAppOrderInput {
  message: string;
  clientName: string;
  clientPhone: string;
}