# lib/api-zod/src/generated/types/whatsAppOrderResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from './order';

export interface WhatsAppOrderResult {
  order: Order;
  parsedMessage: string;
  confidence: number;
}